Organizing and implementing Area 1
I started doing what I mentioned earlier in class. Notice that I added a script just to test an animation things to showcase the idea of the world to the art people. Also, I did a file organization in a way that I think could be more useful to all of us. There's more to come, so please, don't finish the merge yet.