Area 3 starting point
Area 3 is started. I need to do some stuff yet (like add the final part of the puzzle and link to area 4), but it's pretty much on spot to start messing with some stuff. Until Monday, I will do a to do list, so feel free to merge it as you wish since I won't be doing too much in the next couple days.